# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for Bidkan Clean Architecture.",
        "title": "Bidkan Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/v1/admin/notifications/broadcast": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "แอดมินส่งข่าวสาร โปรโมชั่น หรือประกาศสำคัญไปยังผู้ใช้ทุกคนผ่าน MQTT และ Push Notification",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notifications"
                ],
                "summary": "ส่งประกาศแจ้งเตือนหาทุกคน (Admin Only)",
                "parameters": [
                    {
                        "description": "เนื้อหาที่ต้องการประกาศ",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/broadcast.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ส่งประกาศสำเร็จ",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "ไม่มีสิทธิ์ (เฉพาะแอดมินเท่านั้น)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/bikes/{id}/status": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "เปลี่ยนสถานะของรถจักรยาน (เช่น เปลี่ยนเป็น maintenance หรือ unavailable) - เฉพาะ Staff/Admin เท่านั้น",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bikes"
                ],
                "summary": "อัปเดตสถานะรถจักรยาน",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ไอดีรถจักรยาน",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "สถานะใหม่ที่ต้องการเปลี่ยน",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/update_status.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "อัปเดตสำเร็จ",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "ข้อมูลไม่ถูกต้อง",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "ไม่มีสิทธิ์เข้าถึง (ต้องเป็น Staff หรือ Admin)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/coupons/my": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "ดึงรายการคูปองส่วนลดที่ผู้ใช้เก็บสะสมไว้และยังไม่ได้นำไปใช้งาน",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Coupons"
                ],
                "summary": "รายการคูปองของฉัน",
                "responses": {
                    "200": {
                        "description": "รายการคูปองที่ใช้งานได้",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/list_my.CouponInfo"
                            }
                        }
                    },
                    "401": {
                        "description": "ไม่ได้รับอนุญาต",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "เกิดข้อผิดพลาดภายในเซิร์ฟเวอร์",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/coupons/redeem": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "กรอกโค้ดเพื่อรับเงินฟรีเข้า Wallet หรือเก็บโค้ดส่วนลดไว้ใช้หักค่าขี่ในอนาคต",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Coupons"
                ],
                "summary": "ใช้โค้ดคูปอง",
                "parameters": [
                    {
                        "description": "โค้ดคูปองที่ต้องการใช้",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/redeem.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ใช้คูปองสำเร็จ",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "โค้ดไม่ถูกต้อง หรือหมดอายุ หรือถูกใช้ไปแล้ว",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "เกิดข้อผิดพลาดภายในเซิร์ฟเวอร์",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/reports/summary": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "ดึงข้อมูลรายงานสรุปยอดรวม เช่น รายได้รวม จำนวนผู้ใช้ และสถิติการใช้งานรถจักรยาน",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "รายงานสรุปภาพรวมระบบ (Admin Only)",
                "responses": {
                    "200": {
                        "description": "ข้อมูลรายงานสรุป",
                        "schema": {
                            "$ref": "#/definitions/entity.FullReport"
                        }
                    },
                    "403": {
                        "description": "ไม่มีสิทธิ์ (เฉพาะแอดมินเท่านั้น)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "เกิดข้อผิดพลาดภายในเซิร์ฟเวอร์",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/rides/end": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "จบการขี่จักรยาน คำนวณค่าบริการ หักเงินใน Wallet และคืนสถานะรถให้ว่าง",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rides"
                ],
                "summary": "จบการเช่ารถจักรยาน",
                "parameters": [
                    {
                        "description": "ข้อมูลการจบการเช่า",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/end.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "จบการเช่าและชำระเงินสำเร็จ",
                        "schema": {
                            "$ref": "#/definitions/entity.Ride"
                        }
                    }
                }
            }
        },
        "/v1/rides/start": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "เริ่มต้นการขี่จักรยานโดยระบุไอดีรถที่ต้องการ (สแกน QR Code)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rides"
                ],
                "summary": "เริ่มการเช่ารถจักรยาน",
                "parameters": [
                    {
                        "description": "ข้อมูลการเริ่มเช่า",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/start.Request"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "เริ่มการเช่าสำเร็จ",
                        "schema": {
                            "$ref": "#/definitions/entity.Ride"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "broadcast.Request": {
            "type": "object",
            "required": [
                "message",
                "title",
                "type"
            ],
            "properties": {
                "message": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "description": "promo, news, system",
                    "type": "string"
                }
            }
        },
...
        "end.Request": {
            "type": "object",
            "required": [
                "user_id"
            ],
            "properties": {
                "coupon_code": {
                    "description": "เพิ่มฟิลด์เลือกใช้คูปอง (Optional)",
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "entity.BikeUsageStat": {
            "type": "object",
            "properties": {
                "bike_id": {
                    "type": "string"
                },
                "ride_count": {
                    "type": "integer"
                },
                "total_distance_km": {
                    "type": "number"
                },
                "total_revenue": {
                    "type": "string"
                }
            }
        },
        "entity.DailyRevenue": {
            "type": "object",
            "properties": {
                "daily_revenue": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "ride_count": {
                    "type": "integer"
                }
            }
        },
        "entity.FullReport": {
            "type": "object",
            "properties": {
                "bike_stats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.BikeUsageStat"
                    }
                },
                "daily_revenue": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.DailyRevenue"
                    }
                },
                "summary": {
                    "$ref": "#/definitions/entity.SystemSummary"
                }
            }
        },
        "entity.Ride": {
            "type": "object",
            "properties": {
                "bikeID": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "distanceKm": {
                    "type": "number",
                    "format": "float64"
                },
                "endLat": {
                    "type": "number",
                    "format": "float64"
                },
                "endLon": {
                    "type": "number",
                    "format": "float64"
                },
                "endTime": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "startLat": {
                    "type": "number",
                    "format": "float64"
                },
                "startLon": {
                    "type": "number",
                    "format": "float64"
                },
                "startTime": {
                    "type": "string"
                },
                "status": {
                    "description": "ongoing, completed, cancelled",
                    "type": "string"
                },
                "totalFare": {
                    "type": "string"
                },
                "type": {
                    "description": "ride, service",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
        "entity.SystemSummary": {
            "type": "object",
            "properties": {
                "active_rides": {
                    "type": "integer"
                },
                "available_bikes": {
                    "type": "integer"
                },
                "total_bikes": {
                    "type": "integer"
                },
                "total_revenue": {
                    "type": "string"
                },
                "total_users": {
                    "type": "integer"
                }
            }
        },
...
        "list_my.CouponInfo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "collected_at": {
                    "type": "string"
                },
                "expired_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "min_amount": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "user_coupon_id": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "redeem.Request": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "start.Request": {
            "type": "object",
            "required": [
                "bike_id",
                "user_id"
            ],
            "properties": {
                "bike_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "update_status.Request": {
            "type": "object",
            "properties": {
                "bike_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type 'Bearer ' followed by your JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}